+++ tailwind.css
@@ -40,6 +40,7 @@
     --blue-300: 109 155 219;
     --blue-200: 177 202 236;
     --blue-100: 216 228 245;
+    --blue-50: 239 244 251;
   }
 
   .dark {
@@ -71,5 +72,6 @@
     --blue-300: 177 202 236;
     --blue-200: 216 228 245;
     --blue-100: 61 122 207;
+    --blue-50: 239 244 251; /** replace with correct dark color **/
   }
 }
