+++ web-components/src/components/read-more/truncate.ts
@@ -7,13 +7,16 @@ export function truncate(
   str: string,
   maxLength: number,
   restMinLength: number,
+  sentenceBased: boolean,
 ): Texts {
   const regex = /[^\.!\?]+[\.!\?]+/g;
-  const sentences: string[] | null = str.match(regex);
+  let sentences: string[] | null = null;
+  if (sentenceBased) sentences = str.match(regex);
+
   if (
     (maxLength === 0 && restMinLength === 0) ||
     maxLength < restMinLength ||
-    !sentences ||
+    (!sentences && sentenceBased) ||
     str.length < maxLength ||
     str.length < restMinLength
   ) {
@@ -22,10 +25,11 @@ export function truncate(
       rest: '',
     };
   }
+
   let truncated = '',
     tmpTruncated = '';
   let restLength = str.length;
-  if (sentences) {
+  if (sentences && sentenceBased) {
     for (var i: number = 0; i < sentences.length; i++) {
       restLength -= sentences[i].length;
       tmpTruncated += sentences[i];
@@ -34,6 +38,13 @@ export function truncate(
       } else {
         break;
       }
+    }
+  } else {
+    restLength = str.length - maxLength;
+    if (restLength <= restMinLength) {
+      truncated = str;
+    } else {
+      truncated = str.substring(0, maxLength);
     }
   }
   const rest: string = str.substring(truncated.length + 1);
