+++ web-marketplace/src/pages/CreditClassDetails/CreditClassDetailsSimple/CreditClassDetailsSimple.tsx
@@ -187,6 +187,7 @@ const CreditClassDetailsSimple: React.FC<
             </Box>
             {metadata?.['schema:description'] && (
               <ReadMore
+                mobileSize="md"
                 classes={{
                   root: styles.marginBottom,
                   textContainer: styles.textContainer,
